fix: use nouns for weblate badge routes
