+++ services/weblate/weblate-entities.tester.js
